feat: fix payment repository

+++ src/infrastructure/typeorm/entities/user.entity.ts
@@ -5,6 +5,6 @@ export class UserEntity {
   @PrimaryGeneratedColumn('uuid')
   id: string;
 
-  @Column({ type: 'decimal', default: 0 })
+  @Column({ type: 'float', default: 0 })
   balance: number;
 }
